refactor test cases
